removed TestArgsListener and ISuiteListener interface implementation due to the double core methods execution
